Split large files (logged_quantity, base_app) into packages and applied black formatter. No behavior changed and no features added.
