+++ base_app/__init__.py
@@ -1,0 +1,2 @@
+from .base_microscope_app import BaseMicroscopeApp
+from .base_app import BaseApp
