+++ base_app/base_microscope_app_mdi.ui
